Categorias
Imágenes en pagina de categoría, pagina para subir imágenes de categoría, imagen de producto en vista de edición, eliminada sección de listas en vista de productos, imágenes de productos y cambios mínimos

+++ tienda/app/Http/Controllers/ProductsController.php
@@ -6,6 +6,7 @@ use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\File;
 use Illuminate\Http\Request;
 use App\product;
+use App\categories;
 use Auth;
 
 class ProductsController extends Controller
@@ -15,19 +16,16 @@ class ProductsController extends Controller
     {
         $Product = Product::find($brand);
         if($Product){
-            if(($Product->inventory>0)||(Auth::user()&&Auth::user()->type=='admin')){
+            if(($Product->inventory>0)/*||(Auth::user()&&Auth::user()->type=='admin')*/){
                 $title = 'DICES@ | '.$Product->description;
                 return view('Product',['Product'=>$Product], compact('title'));
-            }else{
-                return redirect()->route('home');
             }
-        }else{
-            return redirect()->route('home');
         }
+        return redirect()->route('home');
     }
     public function viewProducts()
     {
-        $Products = DB::table('products')->paginate(20);
+        $Products = DB::table('products')->paginate(1000);
         if($Products){
             return view('Products',['products'=>$Products]);
         }else{
@@ -36,58 +34,36 @@ class ProductsController extends Controller
     }
     public function viewCategory($category)
     {
-        if(Auth::User()){
-            /*if(Auth::User()->type=='admin'){
-                $Products = Product::where('category', $category)->paginate(20);
-            }else*/{
-                $Products = Product::where('category', $category)->where('inventory','>','0')->paginate(20);
+        if(!is_numeric($category))
+        {
+            $categoria = categories::where('subcategoria', $category)->first();
+        }            
+        else
+        {
+            $categoria = categories::where('categoria', $category)->first();
+        }
+        if($categoria)
+        {
+            $Products = Product::where('category', $categoria->subcategoria)->where('inventory','>','0')->paginate(1000);
+            if($Products){
+                $title = 'DICES@ | '.$categoria->titulo;
+                return view('Category',['products'=>$Products,'categoria'=>$categoria], compact('title'));
             }
-        }else{
-            $Products = Product::where('category', $category)->where('inventory','>','0')->paginate(20);
         }
-        if($Products){
-            $title = 'DICES@ | '.$category;
-            return view('Category',['products'=>$Products,'categoria'=>$category], compact('title'));
-        }else{
-            return redirect()->route('home');
-        }
-    }
+        return redirect()->route('home');
 
-    public function viewGamer()
-    {
-        if(Auth::User()){
-            /*if(Auth::User()->type=='admin'){
-                $Products = Product::where('description','LIKE','%GAMER%')->paginate(20);
-            }else*/{
-                $Products = Product::where('description','LIKE','%GAMER%')->where('inventory','>','0')->paginate(20);
-            }
-        }else{
-            $Products = Product::where('description','LIKE','%GAMER%')->where('inventory','>','0')->paginate(20);
-        }
-        if($Products){
-            $title = 'DICES@ | ZONA GAMER';
-            return view('Category',['products'=>$Products,'categoria'=>'Zona Gamer'], compact('title'));
-        }else{
-            return redirect()->route('home');
-        }
     }
 
     public function Search(Request $request){
         $texto=trim($request->get('search'));
-        
-        if(Auth::User()){
-            /*if(Auth::User()->type=='admin'){
-                $resultado = DB::table('products')->where('description','LIKE','%'.$texto.'%')->orwhere('brand','LIKE','%'.$texto.'%')->paginate(10);
-            }else*/{
-                $resultado = DB::table('products')->where('description','LIKE','%'.$texto.'%')->where('inventory','>','0')->orwhere('brand','LIKE','%'.$texto.'%')->paginate(10);
-            }
-        }else{
-            $resultado = DB::table('products')->where('description','LIKE','%'.$texto.'%')->where('inventory','>','0')->orwhere('brand','LIKE','%'.$texto.'%')->paginate(10);
-        }
+        $resultado = DB::table('products')
+                ->where('description','LIKE','%'.$texto.'%')
+                ->where('inventory','>','0')
+                ->orwhere('brand','LIKE','%'.$texto.'%')
+                ->paginate(1000);
         $title = 'DICES@ | '.$texto;
         return view('Search',['resultados'=>$resultado, 'Busqueda'=>$texto], compact('title'));
     }
-
 
     
     /* Vistas de agregar producto */
@@ -146,21 +122,36 @@ class ProductsController extends Controller
                 $Product = Product::find($id);
                 if($Product){
                     return view('Delete',['Product'=>$Product]);
-                }else{
-                    return redirect()->route('home');
                 }
-            }else{
-                return redirect()->route('home');
             }
-        }else{
-            return redirect()->route('home');
         }
+        return redirect()->route('home');
     }
+
     public function Delete($id)
     {
-        $Product = Product::find($id);
-        $Product->delete();
-        return redirect('/home');
+        if(Auth::user()){
+            if(Auth::user()->type==='admin'){
+                $Product = Product::find($id);
+                $Product->delete();
+                return redirect('home');
+            }
+        }
+        return redirect('home');
+    }
+
+    public function viewCatImage($id)
+    {
+        if(Auth::user()){
+            if(Auth::user()->type=='admin'){
+                $categoria = categories::find($id);
+                if($categoria){
+                    $title = 'DICES@ | Imagen para '.$categoria->titulo;
+                    return view('CategoryImage',['categoria'=>$categoria],compact('title'));
+                }
+            }
+        }
+        return redirect('home');
     }
 
     /* Vistas para editar producto */
@@ -170,17 +161,14 @@ class ProductsController extends Controller
             if(Auth::user()->type==='admin'){
                 $Product = Product::find($id);
                 if($Product){
-                    return view('EditProduct',['Product'=>$Product]);
-                }else{
-                    return redirect()->route('home');
+                    $title = 'DICES@ | Editar '.$Product->brand;
+                    return view('EditProduct',['Product'=>$Product],compact('title'));
                 }
-            }else{
-                return redirect()->route('home');
             }
-        }else{
-            return redirect()->route('home');
         }
+        return redirect()->route('home');
     }
+
     public function Edit(Request $request, $id)
     {
         $validateData = $request->validate([
@@ -204,7 +192,11 @@ class ProductsController extends Controller
             $fileName = $data['brand'].'.png';
             $uploadSuccess=$data['image']->move($finalPath,$fileName);
             $product->image = ($finalPath.$fileName);
-        }else{
+        }elseif (file_exists('images/products/'.$data['brand'].'.png'))
+        {
+            $product->image = 'images/products/'.$data['brand'].'.png';
+        }        
+        else{
             $product->image = 'images/sample/productSample.png';
         }
         
@@ -212,5 +204,30 @@ class ProductsController extends Controller
         $product->category = $data['category'];
         $product->save();
         return redirect('product/'.$id);
+    }
+
+    public function catEdit(Request $request, $id)
+    {
+        $validateData = $request->validate([
+            'image' => 'nullable|sometimes|max:10000'
+        ]);
+        $data = request()->all();
+        $category = categories::find($id);
+        
+        if(isset($data['image']))
+        {
+            $file = $data['image'];
+            $finalPath = 'images/category/';
+            if($category['subcategoria']=='CON')
+                $fileName = $category['subcategoria'].'1.png';
+            else
+                $fileName = $category['subcategoria'].'.png';
+            $uploadSuccess=$data['image']->move($finalPath,$fileName);
+            $category->image = ($finalPath.$fileName);
+        }else{
+            $category->image = 'images/sample/categorySample.png';
+        }
+        $category->save();
+        return redirect('category/'.$category->subcategoria);
     }
 }
